move back under scaletest cmd
